+++ src/components/util/CodelabCard/styles.js
@@ -5,6 +5,9 @@ const useStyles = makeStyles((theme) => ({
     margin: "2rem",
     minWidth: "50vw",
   },
+  heading: {
+    fontWeight: "600",
+  },
 }));
 
 export default useStyles;
